refactor: 채팅 함수 분리
