+++ src/components/common/chatbot-search-input/ChatbotSearchInput.tsx
@@ -1,5 +1,10 @@
+import { push, ref, set } from 'firebase/database';
 import React, { FormEvent, useCallback, useRef, useState } from 'react';
 import { picture, upload } from '../../../assets/assets';
+import { database } from '../../../firebase/firebase';
+import useBrandStore from '../../../stores/useBrandStore';
+import useProductStore from '../../../stores/useProductsStore';
+import useUserStore from '../../../stores/useUserStore';
 import {
   chatbotSearchContainer,
   chatbotSearchInput,
@@ -9,18 +14,100 @@ import {
   uploadIconBox,
 } from './chatbotSearchInput.css';
 
-type ChatbotSearchInputProps = {
-  chatCompletionsMutation?: any;
-  onImageUpload?: (file: File) => void;
+import { useMutation } from '@tanstack/react-query';
+import { getDownloadURL, getStorage, ref as storageRef, uploadBytes } from 'firebase/storage';
+import { ImageShoseSearchAPI } from '../../../api/searchRequests';
+import { useChatCompletion } from '../../../hooks/useChatCompletionHook';
+import useChatStore from '../../../stores/useChatStore';
+import { ChatItem } from '../../../types/chatItem';
+
+
+export type Brand = {
+  brand: string;
+  description: string;
+  link: string;
+  thumbnail: string;
 };
 
-const ChatbotSearchInput: React.FC<ChatbotSearchInputProps> = ({ chatCompletionsMutation, onImageUpload }) => {
+
+
+const ChatbotSearchInput = () => {
   const fileInputRef = useRef<HTMLInputElement>(null);
   const inputRef = useRef<HTMLInputElement>(null);
   const [previewImage, setPreviewImage] = useState<string | null>(null);
   const [selectedFile, setSelectedFile] = useState<File | null>(null);
   const [inputValue, setInputValue] = useState<string>('');
   const [isComposing, setIsComposing] = useState<boolean>(false);
+  const { user } = useUserStore();
+  const { setProducts } = useProductStore();
+  const { setBrands } = useBrandStore();
+  const { addChatItem, setCurrentKeywords, } = useChatStore();
+  const { chatCompletionsMutation } = useChatCompletion();
+
+
+
+  const imageSearchMutation = useMutation({
+    mutationFn: async (file: File) => {
+      const storage = getStorage();
+      const imageRef = storageRef(storage, `images/${file.name}`);
+      await uploadBytes(imageRef, file);
+      const imageUrl = await getDownloadURL(imageRef);
+
+      const formData = new FormData();
+      formData.append('image', file);
+      const response = await ImageShoseSearchAPI(formData);
+
+      return { response, imageUrl };
+    },
+    onSuccess: async ({ response, imageUrl }) => {
+      console.log('이미지 검색 성공:', response);
+
+      const chatItemWithoutIds = {
+        userQuestion: '',
+        botResponse: '이미지 검색 결과입니다.',
+        products: response.data.products || null,
+        brands: response.data.brands || null,
+        keywords: '이미지 검색',
+        imageUrl: imageUrl,
+        timestamp: new Date().toISOString(),
+      };
+
+      const shareId = await saveSharedChatHistory(chatItemWithoutIds);
+
+      const newChatItem: ChatItem = {
+        id: push(ref(database, 'chatHistory')).key || '',
+        shareId,
+        ...chatItemWithoutIds,
+      };
+
+      if (user?.uid) {
+        push(ref(database, `chatHistory/${user.uid}`), newChatItem);
+      }
+      setProducts(response.data.products);
+      setBrands(response.data.brands);
+      setCurrentKeywords('이미지 검색');
+
+      addChatItem(newChatItem);
+    },
+    onError: error => {
+      console.error('이미지 검색 에러:', error);
+    },
+  });
+
+  const saveSharedChatHistory = async (chatItem: Omit<ChatItem, 'id' | 'shareId'>): Promise<string> => {
+    const shareId = push(ref(database, 'sharedChatHistory')).key;
+    if (shareId) {
+      await set(ref(database, `sharedChatHistory/${shareId}`), chatItem);
+      console.log('공유용 채팅 히스토리 저장 성공:', shareId);
+      return shareId;
+    }
+    throw new Error('Failed to generate shareId');
+  };
+
+  const handleImageUpload = (file: File) => {
+    imageSearchMutation.mutate(file);
+  };
+
 
   const debounce = (func: Function, delay: number) => {
     let timeoutId: NodeJS.Timeout;
@@ -48,7 +135,7 @@ const ChatbotSearchInput: React.FC<ChatbotSearchInputProps> = ({ chatCompletions
   const handleSubmit = (e?: FormEvent) => {
     e?.preventDefault();
     if (selectedFile) {
-      onImageUpload?.(selectedFile);
+      handleImageUpload?.(selectedFile);
       setSelectedFile(null);
       setPreviewImage(null);
       if (fileInputRef.current) {
